+++ app/src/main/java/jp/aoyama/mki/thermometer/domain/repository/BluetoothDeviceScanner.kt
@@ -1,6 +1,7 @@
-package jp.aoyama.mki.thermometer.infrastructure.bluetooth
+package jp.aoyama.mki.thermometer.domain.repository
 
 import androidx.lifecycle.LiveData
+import jp.aoyama.mki.thermometer.domain.models.BluetoothDeviceData
 
 /**
  * 付近のBluetooth端末を検索し、[BluetoothDeviceScanner.devicesLiveData]により配信する。
