+++ app/src/main/java/jp/aoyama/mki/thermometer/domain/service/UserService.kt
@@ -5,7 +5,7 @@ import jp.aoyama.mki.thermometer.domain.models.Grade
 import jp.aoyama.mki.thermometer.domain.models.User
 import jp.aoyama.mki.thermometer.domain.models.UserEntity
 import jp.aoyama.mki.thermometer.domain.repository.UserRepository
-import jp.aoyama.mki.thermometer.infrastructure.user.api.UserApiRepository
+import jp.aoyama.mki.thermometer.infrastructure.api.user.UserApiRepository
 import jp.aoyama.mki.thermometer.view.models.UserData
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
