+++ app/src/main/java/jp/aoyama/mki/thermometer/infrastructure/api/user/models/EditDeviceRequest.kt
@@ -1,0 +1,5 @@
+package jp.aoyama.mki.thermometer.infrastructure.api.user.models
+
+data class EditDeviceRequest(
+    val devices: List<String>
+)
