+++ app/src/main/java/jp/aoyama/mki/thermometer/infrastructure/api/user/models/EditGradeRequest.kt
@@ -1,0 +1,5 @@
+package jp.aoyama.mki.thermometer.infrastructure.api.user.models
+
+data class EditGradeRequest(
+    val grade: String?
+)
