+++ app/src/main/java/jp/aoyama/mki/thermometer/infrastructure/csv/temperature/LocalFileTemperatureRepository.kt
@@ -1,4 +1,4 @@
-package jp.aoyama.mki.thermometer.infrastructure.temperature.file
+package jp.aoyama.mki.thermometer.infrastructure.csv.temperature
 
 import android.content.Context
 import android.util.Log
