+++ app/src/main/java/jp/aoyama/mki/thermometer/infrastructure/csv/user/UserCSVUtil.kt
@@ -1,4 +1,4 @@
-package jp.aoyama.mki.thermometer.infrastructure.user
+package jp.aoyama.mki.thermometer.infrastructure.csv.user
 
 import android.content.Context
 import android.net.Uri
