--- app/src/main/java/jp/aoyama/mki/thermometer/infrastructure/user/api/models/EditDeviceRequest.kt
@@ -1,5 +1,0 @@
-package jp.aoyama.mki.thermometer.infrastructure.user.api.models
-
-data class EditDeviceRequest(
-    val devices: List<String>
-)
