--- app/src/main/java/jp/aoyama/mki/thermometer/infrastructure/user/api/models/EditGradeRequest.kt
@@ -1,5 +1,0 @@
-package jp.aoyama.mki.thermometer.infrastructure.user.api.models
-
-data class EditGradeRequest(
-    val grade: String?
-)
