+++ app/src/main/java/jp/aoyama/mki/thermometer/view/user/fragments/AddBluetoothDeviceFragment.kt
@@ -13,8 +13,8 @@ import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
 import jp.aoyama.mki.thermometer.databinding.FeagmentAddBluetoothDeviceBinding
 import jp.aoyama.mki.thermometer.domain.models.BluetoothData
-import jp.aoyama.mki.thermometer.infrastructure.bluetooth.BluetoothDeviceScanner
-import jp.aoyama.mki.thermometer.infrastructure.bluetooth.android.BluetoothDiscoveryDeviceScanner
+import jp.aoyama.mki.thermometer.domain.repository.BluetoothDeviceScanner
+import jp.aoyama.mki.thermometer.infrastructure.android.bluetooth.BluetoothDiscoveryDeviceScanner
 import jp.aoyama.mki.thermometer.view.bluetooth.list.BluetoothListAdapter
 import jp.aoyama.mki.thermometer.view.bluetooth.list.BluetoothViewHolder
 import jp.aoyama.mki.thermometer.view.user.viewmodels.UserViewModel
